requirements.txt
pyodbc>=5.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

